SF-33 Moved validators around
