+++ frontend/src/components/validators.js
@@ -1,7 +1,5 @@
 import PropTypes from 'prop-types'
 
-import PreviewCompatibleImage from './PreviewCompatibleImage'
-
 // These are validators that can be re-used to validate events and lists of events
 export const eventPropTypeValidator = PropTypes.shape({
   frontmatter: PropTypes.shape({
@@ -16,22 +14,35 @@ export const eventPropTypeValidator = PropTypes.shape({
   })
 })
 
-export const newsArticleValidator = PropTypes.shape({
-  node: PropTypes.shape({
-    frontmatter: PropTypes.shape({
-      title: PropTypes.string,
-      image: PreviewCompatibleImage.propTypes.imageInfo
-    }),
-    fields: PropTypes.shape({
-      slug: PropTypes.string
-    })
-  })
-})
-
 export const gatsbyImageSharpFluidValidator = PropTypes.shape({
   base64: PropTypes.string.isRequired,
   aspectRatio: PropTypes.number.isRequired,
   src: PropTypes.string.isRequired,
   srcSet: PropTypes.string.isRequired,
   sizes: PropTypes.string.isRequired
+})
+
+export const childImageSharpValidator = PropTypes.shape({
+  fluid: gatsbyImageSharpFluidValidator
+})
+
+export const previewCompatibleImageValidator = PropTypes.shape({
+  alt: PropTypes.string,
+  image: PropTypes.oneOfType([
+    PropTypes.string,
+    childImageSharpValidator
+  ]),
+  value: PropTypes.string
+})
+
+export const newsArticleValidator = PropTypes.shape({
+  node: PropTypes.shape({
+    frontmatter: PropTypes.shape({
+      title: PropTypes.string,
+      image: previewCompatibleImageValidator
+    }),
+    fields: PropTypes.shape({
+      slug: PropTypes.string
+    })
+  })
 })
